Convert email preferences to es6
